rearrange to library structure
