+++ Cargo.toml
@@ -3,7 +3,13 @@ name = "boring-derive"
 version = "0.1.0"
 edition = "2021"
 
+[lib]
+proc-macro = true
+
 [dependencies]
+proc-macro2 = "1.0.86"
 quote = "1.0.36"
 syn = "2.0.67"
-boring_derive_macro = {path = "../boring-derive/boring_derive_macro/"}
+
+[dev-dependencies]
+trybuild = "1.0.96"
