--- boring_derive_macro/.gitignore
@@ -1,1 +1,0 @@
-/target
