--- boring_derive_macro/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "boring_derive_macro"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-proc-macro = true
-
-[dependencies]
-proc-macro2 = "1.0.86"
-quote = "1.0.36"
-syn = "2.0.67"
