+++ src/core/symbol.rs
