+++ tests/test.rs
@@ -1,0 +1,5 @@
+#[test]
+fn test() {
+    let t = trybuild::TestCases::new();
+    t.pass("tests/builder/example.rs")
+}
